+++ src/TextShow/FromStringTextShow.hs
@@ -4,12 +4,11 @@
 {-# LANGUAGE DeriveFunctor              #-}
 {-# LANGUAGE DeriveTraversable          #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
+{-# LANGUAGE TemplateHaskell            #-}
+{-# LANGUAGE TypeFamilies               #-}
 
 #if __GLASGOW_HASKELL__ >= 702
 {-# LANGUAGE DeriveGeneric              #-}
-#else
-{-# LANGUAGE TemplateHaskell            #-}
-{-# LANGUAGE TypeFamilies               #-}
 #endif
 
 #if __GLASGOW_HASKELL__ >= 708
@@ -32,13 +31,15 @@ module TextShow.FromStringTextShow (FromStringShow(..), FromTextShow(..)) where
 
 import           Data.Data (Data, Typeable)
 
+#if __GLASGOW_HASKELL__ < 706
+import qualified Generics.Deriving.TH as Generics
+#endif
+
 #if __GLASGOW_HASKELL__ >= 702
 import           GHC.Generics (Generic)
 # if __GLASGOW_HASKELL__ >= 706
 import           GHC.Generics (Generic1)
 # endif
-#else
-import qualified Generics.Deriving.TH as Generics (deriveAll)
 #endif
 
 import           Prelude ()
@@ -137,7 +138,14 @@ instance TextShow1 FromTextShow where
 
 -------------------------------------------------------------------------------
 
+#if __GLASGOW_HASKELL__ < 706
+$(Generics.deriveMeta           ''FromStringShow)
+$(Generics.deriveRepresentable1 ''FromStringShow)
+$(Generics.deriveMeta           ''FromTextShow)
+$(Generics.deriveRepresentable1 ''FromTextShow)
+#endif
+
 #if __GLASGOW_HASKELL__ < 702
-$(Generics.deriveAll ''FromStringShow)
-$(Generics.deriveAll ''FromTextShow)
+$(Generics.deriveRepresentable0 ''FromStringShow)
+$(Generics.deriveRepresentable0 ''FromTextShow)
 #endif