+++ tests/Spec/Derived/MagicHashSpec.hs
@@ -15,7 +15,7 @@ module Spec.Derived.MagicHashSpec (main, spec) where
 
 import Derived.MagicHash
 
-import Spec.Utils (prop_matchesTextShow)
+import Spec.Utils (prop_matchesTextShow2, prop_genericTextShow, prop_genericTextShow1)
 
 import Test.Hspec (Spec, describe, hspec, parallel)
 import Test.Hspec.QuickCheck (prop)
@@ -25,9 +25,13 @@ main = hspec spec
 
 spec :: Spec
 spec = parallel $ do
-    describe "TyCon#" $
-        prop "TextShow instance" (prop_matchesTextShow :: Int -> TyCon# -> Bool)
+    describe "TyCon# Int Int" $ do
+        prop "TextShow2 instance" (prop_matchesTextShow2 :: Int -> TyCon# Int Int -> Bool)
+        prop "generic TextShow"   (prop_genericTextShow  :: Int -> TyCon# Int Int -> Bool)
+        prop "generic TextShow1"  (prop_genericTextShow1 :: Int -> TyCon# Int Int -> Bool)
 #if MIN_VERSION_template_haskell(2,7,0)
-    describe "TyFamily#" $
-        prop "TextShow instance" (prop_matchesTextShow :: Int -> TyFamily# -> Bool)
+    describe "TyFamily# Int Int" $ do
+        prop "TextShow2 instance" (prop_matchesTextShow2 :: Int -> TyFamily# Int Int -> Bool)
+        prop "generic TextShow"   (prop_genericTextShow  :: Int -> TyFamily# Int Int -> Bool)
+        prop "generic TextShow1"  (prop_genericTextShow1 :: Int -> TyFamily# Int Int -> Bool)
 #endif