+++ tests/Spec/Derived/RecordsSpec.hs
@@ -15,9 +15,6 @@ module Spec.Derived.RecordsSpec (main, spec) where
 import Derived.Records
 
 import Spec.Utils (prop_matchesTextShow2, prop_genericTextShow, prop_genericTextShow1)
-#if MIN_VERSION_template_haskell(2,7,0)
-import Spec.Utils (prop_genericTextShow')
-#endif
 
 import Test.Hspec (Spec, describe, hspec, parallel)
 import Test.Hspec.QuickCheck (prop)
@@ -34,6 +31,6 @@ spec = parallel $ do
 #if MIN_VERSION_template_haskell(2,7,0)
     describe "TyFamily Int Int" $ do
         prop "TextShow2 instance" (prop_matchesTextShow2 :: Int -> TyFamily Int Int -> Bool)
-        prop "generic TextShow"   (prop_genericTextShow' :: Int -> TyFamily Int Int -> Bool)
+        prop "generic TextShow"   (prop_genericTextShow  :: Int -> TyFamily Int Int -> Bool)
         prop "generic TextShow1"  (prop_genericTextShow1 :: Int -> TyFamily Int Int -> Bool)
 #endif